afs-client-accessd: add signal to force export
Currently, the sqlite files are exported to the central host at
midnight. As a result, the admin has to wait until the end of the day
to check if the export mode in question is working as expected. To avoid
this inconvenience, use SIGUSR1 to immediately export the sqlite files
to the central host. This patch should not be used in production.